mantapay config works
