+++ manta-accounting/src/transfer/utxo/protocol.rs
@@ -1132,8 +1132,8 @@ where
     C: Configuration<Bool = bool>,
     C::AssetId: Clone + Default,
     C::AssetValue: Clone + Default,
-    IncomingBaseRandomness<C>: Clone,
-    IncomingRandomness<C>: Sample,
+    C::Scalar: Sample,
+    IncomingBaseRandomness<C>: Clone + Sample,
     UtxoCommitmentRandomness<C>: Sample,
 {
     #[inline]
@@ -1150,7 +1150,10 @@ where
         let address_partition = self.address_partition_function.partition(&address);
         let secret = MintSecret::<C>::new(
             address.receiving_key,
-            rng.gen(),
+            Randomness {
+                ephemeral_secret_key: rng.gen(),
+                randomness: rng.gen(),
+            }, // FIXME: manta-pay doesn't compile when I simply use rng.gen() here.
             IncomingPlaintext::new(rng.gen(), associated_data.secret(&asset)),
         );
         let utxo_commitment = self.base.utxo_commitment_scheme.commit(
