+++ manta-pay/src/lib.rs
@@ -33,9 +33,9 @@ pub mod config;
 #[cfg_attr(doc_cfg, doc(cfg(feature = "key")))]
 pub mod key;
 
-#[cfg(all(feature = "groth16", feature = "test"))]
-#[cfg_attr(doc_cfg, doc(cfg(all(feature = "groth16", feature = "test"))))]
-pub mod parameters;
+// #[cfg(all(feature = "groth16", feature = "test"))]
+// #[cfg_attr(doc_cfg, doc(cfg(all(feature = "groth16", feature = "test"))))]
+// pub mod parameters;
 
 // #[cfg(feature = "groth16")]
 // #[cfg_attr(doc_cfg, doc(cfg(feature = "groth16")))]
